pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "asl-talk"
version = "0.1.0"
description = "ASL recognition system"
requires-python = ">=3.9"

dependencies = [
    "opencv-python",
    "mediapipe",
    "numpy",
    "pyttsx3",
    "torch",
    "pyyaml"
]

[tool.setuptools.packages.find]
where = ["src"]